Module Octez_smart_rollup_node_PsQuebec.Dal_slots_tracker

The rollup node keeps the list of dal slots for each block it needs to process. This is to determine whether the inbox for a given block will need to be retrieved from the block operations, or from the data availability layer after lag levels have passed and the slot for the block has been declared available.

The state of slots per block is persistent.

val process_head : Octez_smart_rollup_node.Node_context.rw -> Octez_smart_rollup_node.Layer1.head -> unit Tezos_base.TzPervasives.tzresult Lwt.t

process_head node_ctxt head performs the following operations:

it reads the attestations for headers published attestation_lag levels preceding head from the block metadata, determines which ones the rollup node will download, and stores the results in Store.Dal_confirmed_slots.