Module Octez_smart_rollup_node_PsQuebec.Reveals

This module provides basic support for reveals.

The rollup can ask for data being the reveal of some hash. This allows transferring data directly to the rollup without going through the L1 inbox.

Data length must be under 4KB to be refutable in a single L1 operation.

Data must be made available by off-chain mechanisms: it is the responsibility of the rollup kernel to make sure that the reveal data is available: otherwise, there is a potential safety issue.

For the moment, the support is basic and mostly manual as the operator needs to explicitly import a file in the rollup node data directoy to enable the rollup node to answer reveal requests.

type source =

| String of string
(*
A string containing the whole data
*)

| File of string
(*
A file name whose associated file contains the whole data
*)

Source of data

val get : pre_images_endpoint:Uri.t option -> data_dir:string -> pvm_kind:Octez_smart_rollup.Kind.t -> Tezos_protocol_021_PsQuebec.Protocol.Sc_rollup_reveal_hash.t -> string Tezos_base.TzPervasives.tzresult Lwt.t

get ~pre_images_endpoint ~data_dir ~pvm_name ~hash retrieves the data associated with the reveal hash hash from disk. If the data is not already on disk, from an HTTP service at pre_images_endpoint. May fail with:

Wrong_hash {found; expected} where expected = hash, and found <> hash, if the data is retrieved and hashes to the wrong hash found,

Could_not_open_preimage_file filename if the function tries to retrieve the data from filename, but it cannot read the contents of the file.

Could_not_encode_raw_data if the data is too large (more than 4kB) to be revealed.

val proto_hash_to_bytes : Tezos_protocol_021_PsQuebec.Protocol.Sc_rollup_reveal_hash.t -> bytes

Conversion from protocol reveal hash to protocol agnostic hash.